+++ mdObject/src/classes/Problem.ts
@@ -1,8 +1,21 @@
-import { ObjectStatus } from '../enums';
+import { ObjectState, ObjectStatus } from '../enums';
 import { Condition } from '../fhir';
+import { EmrMel } from './EmrMel';
+import { System } from './System';
 
 export class Problem {
     status: ObjectStatus = ObjectStatus.Unchanged;
+    state: ObjectState = ObjectState.None;
+
+    // The assessment values:
+    // New or N
+    // Improved or I
+    // Unchanged or U
+    // Deteriorated or D
+    // Comment only or C
+    // Note. Codes that were supported in previous releases may also appear as Better (B), Same (S), or Worse (W).
+    assessment: string = 'N';
+    note: string = '';
 
     problemId: string = '';
     type: string = '';
@@ -14,6 +27,49 @@ export class Problem {
     stopReason: string = '';
     codeIcd10: string = '';
     lastModifiedDate: string = '';
+
+    save = async (mel: EmrMel) => {
+        switch (this.state) {
+            case ObjectState.Add: {
+                let code: string = await mel.melFunc('{MEL_ADD_PROBLEM("' + this.toAddString() + '")}');
+
+                if (code !== '') {
+                    let error = 'Problem.save error. Code is ' + code;
+                    console.error(error);
+                    throw new Error('Problem not saved. ' + error);
+                }
+
+                this.status = ObjectStatus.Added;
+                this.state = ObjectState.None;
+                break;
+            }
+        }
+    }
+
+    // TODO: Add unit test for:
+    // 1. codeIcd9 = empty && codeIcd10 = ICD10-A (Result 'ICD10-A')
+    // 2. codeIcd9 = ICD-B && codeIcd10 = ICD10-A (Result 'ICD-B|ICD10-A')
+    // 3. codeIcd9 = empty && codeIcd10 = empty (result '')
+    // 4. codeIcd9 = ICD-B && codeIcd10 = empty (Result 'ICD-B')
+    private get code()  {
+        let code = '';
+        code = (this.codeIcd9) ? this.codeIcd9 : code;
+        code = (this.codeIcd10) ? (code) ? '|' + this.codeIcd10 : this.codeIcd10 : code;
+
+        return code;
+}
+
+    private toAddString = (): string => {
+        return this.type + '","' +
+            this.description + '","' +
+            this.code + '","' +
+            (this.onsetDate ? System.dateToString(this.onsetDate) : '') + '","' +
+            (this.stopDate ? System.dateToString(this.stopDate) : '') + '","' +
+            this.comment + '",' +
+            this.assessment + ',"' +
+            this.note;
+    }
+
 
     // Usage:
     // let problem: Problem = Problem.fromFhir(condition);