+++ mdObject/src/classes/EmrObject.ts
@@ -1,6 +1,10 @@
 import { EmrMel } from "./EmrMel";
+import { ObjectState, ObjectStatus } from '../enums';
 
 export abstract class EmrObject<T> {
+    status: ObjectStatus = ObjectStatus.Unchanged;
+    state: ObjectState = ObjectState.None;
+
     // All base properties MUST have an initial value!!! 
     protected objectSetup(base?: T, item?: T) {
         if (item && base) {