+++ mdObject/src/fhir/Meta.ts
@@ -1,0 +1,4 @@
+export class Meta {
+    versionId?: string;
+    lastUpdated?: string;
+}