+++ mdObject/src/fhir/Reference.ts
@@ -1,0 +1,4 @@
+export class Reference {
+    reference?: string; // C? Relative, internal or absolute URL reference
+    display?: string;   // Text alternative for the resource
+}