+++ mdObject/src/fhir/Resource.ts
@@ -1,7 +1,9 @@
+import { Meta } from "./Meta";
+
 export class Resource {
-    resourceType: string;
-    id: string; // Logical id of this artifact
-//--not implemented    "meta": { Meta }, // Metadata about the resource
+    resourceType?: string;
+    id?: string; // Logical id of this artifact
+    meta?: Meta; // Metadata about the resource
 //--not implemented    "implicitRules": "<uri>", // A set of rules under which this content was created
-    language: string; // Language of the resource content 
+    language?: string; // Language of the resource content 
 }