+++ mdObject/src/fhir/types/Converter.ts
@@ -1,0 +1,8 @@
+import { dateTime } from "./dateTime";
+
+export class Converter {
+
+    static dateTimeToDate(source: dateTime): Date {
+        return new Date(source);
+    }
+}