+++ mdObject/src/fhir/types/dateTime.ts
@@ -1,0 +1,2 @@
+export type dateTime = string & {
+}