+++ mdObject/src/fhir/Search.ts
@@ -1,0 +1,4 @@
+export class Search {
+    mode?: string;
+    score?: number;
+}