+++ mdObject/src/fhir/codes/BundleTypeCode.ts
@@ -1,0 +1,4 @@
+export type BundleTypeCode =
+    'document' | 'message' | 'transaction' |
+    'transaction-response' | 'batch' | 'batch-response' |
+    'history' | 'searchset' | 'collection';