+++ mdObject/src/public.ts
@@ -15,3 +15,4 @@ import { MdObject } from './classes';
 export * from './classes';
 export * from './enums';
 export * from './interfaces';
+export * as Fhir from './fhir'