+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 The * indicates changes that break backward compatibility. See https://docs.npmjs.com/about-semantic-versioning for more about product versioning.  
 ## Release 2.0.0-alpha.2.4
 	Added implementation for Problem.save()
+	New abstract EmrObject<T> added.
+	Problem extends EmrObject<Problem>
+	Added sample for FHIR Conditions.
+	Added FHIR implementation for types.
 
 ## Release 2.0.0-alpha.2.3
 	Added export EmrContents class.