+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Note
 The * indicates changes that break backward compatibility. See https://docs.npmjs.com/about-semantic-versioning for more about product versioning.  
+## Release 2.0.0-alpha.2.5
+
 ## Release 2.0.0-alpha.2.4
 	Added implementation for Problem.save()
 	New abstract EmrObject<T> added.